Fix CI: Apply Foundry code formatting standards
- Run 'forge fmt' to fix all formatting issues
- Ensures CI passes the 'forge fmt --check' step
- No functional changes, only whitespace/formatting adjustments
- All tests and functionality remain unchanged

+++ script/TestEndToEnd.s.sol
@@ -23,33 +23,33 @@ contract TestEndToEnd is Script {
     function run() external {
         uint256 deployerPrivateKey = vm.envUint("PRIVATE_KEY");
         address deployer = vm.addr(deployerPrivateKey);
-        
+
         // Get deployed addresses from environment
         address hookAddress = vm.envAddress("HOOK_ADDRESS");
         address poolManagerAddress = vm.envAddress("POOL_MANAGER_ADDRESS");
-        
+
         console.log("=== End-to-End IL Insurance Hook Test ===");
         console.log("Deployer:", deployer);
         console.log("Hook Address:", hookAddress);
         console.log("PoolManager Address:", poolManagerAddress);
-        
+
         vm.startBroadcast(deployerPrivateKey);
-        
+
         // Connect to deployed contracts
         ILInsuranceHook hook = ILInsuranceHook(hookAddress);
         IPoolManager poolManager = IPoolManager(poolManagerAddress);
-        
+
         // Deploy test routers
         PoolModifyLiquidityTest liquidityRouter = new PoolModifyLiquidityTest(poolManager);
         PoolSwapTest swapRouter = new PoolSwapTest(poolManager);
-        
+
         console.log("Liquidity Router:", address(liquidityRouter));
         console.log("Swap Router:", address(swapRouter));
-        
+
         // Use mock tokens for testing (you'd use real tokens on testnet)
         Currency currency0 = Currency.wrap(address(0x1000));
         Currency currency1 = Currency.wrap(address(0x2000));
-        
+
         // Create pool key
         PoolKey memory poolKey = PoolKey({
             currency0: currency0,
@@ -58,12 +58,12 @@ contract TestEndToEnd is Script {
             tickSpacing: 60,
             hooks: hook
         });
-        
+
         PoolId poolId = poolKey.toId();
-        
+
         console.log("\n=== Step 1: Initialize Pool ===");
         uint160 initialSqrtPriceX96 = TickMath.getSqrtPriceAtTick(0); // 1:1 price
-        
+
         try poolManager.initialize(poolKey, initialSqrtPriceX96) {
             console.log("SUCCESS: Pool initialized successfully");
             console.log("Pool ID:", uint256(PoolId.unwrap(poolId)));
@@ -73,7 +73,7 @@ contract TestEndToEnd is Script {
             vm.stopBroadcast();
             return;
         }
-        
+
         console.log("\n=== Step 2: Add Liquidity ===");
         IPoolManager.ModifyLiquidityParams memory addParams = IPoolManager.ModifyLiquidityParams({
             tickLower: -60,
@@ -81,7 +81,7 @@ contract TestEndToEnd is Script {
             liquidityDelta: 1000e18,
             salt: bytes32(0)
         });
-        
+
         // Mock token transfers for testing
         vm.mockCall(
             Currency.unwrap(currency0),
@@ -93,45 +93,43 @@ contract TestEndToEnd is Script {
             abi.encodeWithSignature("transferFrom(address,address,uint256)"),
             abi.encode(true)
         );
-        
+
         try liquidityRouter.modifyLiquidity(poolKey, addParams, bytes("")) {
             console.log("SUCCESS: Liquidity added successfully");
             console.log("Expected: DebugAddLiquidity event should be emitted");
         } catch Error(string memory reason) {
             console.log("FAILED: Add liquidity failed:", reason);
         }
-        
+
         console.log("\n=== Step 3: Perform Swaps (Collect Premiums) ===");
         IPoolManager.SwapParams memory swapParams = IPoolManager.SwapParams({
             zeroForOne: true,
             amountSpecified: -100e18,
             sqrtPriceLimitX96: TickMath.MIN_SQRT_PRICE + 1
         });
-        
+
         // Perform multiple swaps to collect premiums
-        for (uint i = 0; i < 3; i++) {
-            try swapRouter.swap(poolKey, swapParams, PoolSwapTest.TestSettings({
-                takeClaims: false,
-                settleUsingBurn: false
-            }), bytes("")) {
+        for (uint256 i = 0; i < 3; i++) {
+            try swapRouter.swap(
+                poolKey, swapParams, PoolSwapTest.TestSettings({takeClaims: false, settleUsingBurn: false}), bytes("")
+            ) {
                 console.log("SUCCESS: Swap", i + 1, "completed - premium collected");
             } catch Error(string memory reason) {
                 console.log("FAILED: Swap", i + 1, "failed:", reason);
             }
-            
+
             // Alternate swap direction
             swapParams.zeroForOne = !swapParams.zeroForOne;
-            swapParams.sqrtPriceLimitX96 = swapParams.zeroForOne ? 
-                TickMath.MIN_SQRT_PRICE + 1 : 
-                TickMath.MAX_SQRT_PRICE - 1;
+            swapParams.sqrtPriceLimitX96 =
+                swapParams.zeroForOne ? TickMath.MIN_SQRT_PRICE + 1 : TickMath.MAX_SQRT_PRICE - 1;
         }
-        
+
         console.log("\n=== Step 4: Simulate Price Change & Remove Liquidity ===");
         console.log("In a real test, you would:");
         console.log("1. Perform large swaps to move price significantly");
         console.log("2. Remove liquidity to trigger IL calculation");
         console.log("3. Verify compensation is paid if IL > threshold");
-        
+
         // Remove liquidity (this should trigger IL calculation)
         IPoolManager.ModifyLiquidityParams memory removeParams = IPoolManager.ModifyLiquidityParams({
             tickLower: -60,
@@ -139,7 +137,7 @@ contract TestEndToEnd is Script {
             liquidityDelta: -1000e18,
             salt: bytes32(0)
         });
-        
+
         try liquidityRouter.modifyLiquidity(poolKey, removeParams, bytes("")) {
             console.log("SUCCESS: Liquidity removed successfully");
             console.log("Expected: DebugRemoveLiquidity event should be emitted");
@@ -147,21 +145,21 @@ contract TestEndToEnd is Script {
         } catch Error(string memory reason) {
             console.log("FAILED: Remove liquidity failed:", reason);
         }
-        
+
         console.log("\n=== Test Summary ===");
         console.log("SUCCESS: Hook deployment verified");
         console.log("SUCCESS: Pool initialization tested");
         console.log("SUCCESS: Liquidity operations tested");
         console.log("SUCCESS: Swap operations tested");
         console.log("SUCCESS: End-to-end flow validated");
-        
+
         console.log("\n=== Next Steps ===");
         console.log("1. Monitor events on Etherscan for DebugAddLiquidity/DebugRemoveLiquidity");
         console.log("2. Test with real tokens and larger amounts");
         console.log("3. Validate gas costs and optimize if needed");
         console.log("4. Test edge cases (large IL, insufficient insurance pool, etc.)");
         console.log("5. Deploy to mainnet when ready");
-        
+
         vm.stopBroadcast();
     }
 }
